+++ src/main/java/system/customer/repository/ICustomerRepository.java
@@ -5,5 +5,5 @@ import system.customer.model.CustomerModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ICustomerRepository extends JpaRepository<CustomerModel, Integer> {
-    CustomerModel findByDocument(Integer document);
+    CustomerModel findByDocument(long document);
 }
